requirements.txt
Flask==2.1.0
Flask-SocketIO==5.1.1
python-socketio==5.5.0
Pillow==9.1.0
numpy==1.22.3
Werkzeug==2.2.2